refactor: adopt react as primary frontend
